+++ asterisk/agi/src/Agi/Action/CallForwardAction.php
@@ -3,6 +3,7 @@ namespace Agi\Action;
 
 use Agi\Agents\ResidentialAgent;
 use Agi\Agents\UserAgent;
+use Agi\Agents\FriendAgent;
 use Agi\ChannelInfo;
 use Agi\Wrapper;
 use Ivoz\Provider\Domain\Model\CallForwardSetting\CallForwardSettingInterface;
@@ -80,15 +81,20 @@ class CallForwardAction
             $this->agi->setVariable("_USERID", $forwarder->getId());
         } else {
             $forwarder = $this->cfw->getResidentialDevice();
+            if ($forwarder) {
+                $caller = new ResidentialAgent($this->agi, $forwarder);
+                $this->agi->setVariable("_RESIDENTIALDEVICEID", $forwarder->getId());
+            } else {
+                $forwarder = $this->cfw->getFriend();
+                if (!$forwarder) {
+                    // Cfw without owner. This should not happen.
+                    $this->agi->error("Call forward without owner. Check configuration.");
+                    return;
+                }
 
-            if (!$forwarder) {
-                // Cfw without owner. This should not happen.
-                $this->agi->error("Call forward without owner. Check configuration.");
-                return;
+                $caller = new FriendAgent($this->agi, $forwarder);
+                $this->agi->setVariable("_FRIENDID", $forwarder->getId());
             }
-
-            $caller = new ResidentialAgent($this->agi, $forwarder);
-            $this->agi->setVariable("_RESIDENTIALDEVICEID", $forwarder->getId());
         }
 
         // Set the new caller
@@ -128,13 +134,13 @@ class CallForwardAction
         $this->agi->setRedirecting('from-num', $caller->getExtensionNumber());
 
         // Set voicemail destination
-        $isUserCallForward = $this->cfw->getUser() !== null;
-        if ($isUserCallForward) {
-            $this->routerAction
-                ->setRouteVoicemailUser($this->cfw->getVoiceMailUser(), true);
-        } else {
+        $isResidentialCallForward = $this->cfw->getResidentialDevice() !== null;
+        if ($isResidentialCallForward) {
             $this->routerAction
                 ->setRouteVoicemailResidential($this->cfw->getResidentialDevice(), true);
+        } else {
+            $this->routerAction
+                ->setRouteVoicemailUser($this->cfw->getVoiceMailUser(), true);
         }
 
         // Route based on configured type