+++ library/DataFixtures/ORM/KamTrunksCdr.php
@@ -46,6 +46,35 @@ class KamTrunksCdr extends Fixture implements DependentFixtureInterface
         $this->sanitizeEntityValues($item1);
         $manager->persist($item1);
 
+        /////////////////////////
+        ///
+        ////////////////////////
+
+        /** @var TrunksCdrInterface $item2 */
+        $item2 = $this->createEntityInstance(TrunksCdr::class);
+        (function () use ($fixture) {
+            $this->setStartTime(new \DateTime('2021-11-22 16:54:49'));
+            $this->setEndTime(new \DateTime('2021-11-22 16:54:54'));
+            $this->setDuration(4.765);
+            $this->setDirection('outbound');
+
+            $this->setCaller('+3494696823899');
+            $this->setCallee('+34676238611');
+            $this->setCallid('017cc7c8-eb38-4bbd-9318-524a274f7000');
+            $this->setCallidHash('2789d532');
+            $this->setXcallid('9297bdde-309cd48f@10.10.1.123');
+            $this->setParsed(1);
+            $this->setCgrid('5a364b1fe35e00fb2ac1923b43f84eeb78400e03');
+            $this->setParserScheduledAt(new \DateTime('2018-11-22 16:54:54'));
+            $this->setBrand($fixture->getReference('_reference_ProviderBrand1'));
+            $this->setCompany($fixture->getReference('_reference_ProviderCompany1'));
+            $this->setCarrier($fixture->getReference('_reference_ProviderCarrier1'));
+        })->call($item2);
+
+        $this->addReference('_reference_KamTrunksCdr2', $item2);
+        $this->sanitizeEntityValues($item2);
+        $manager->persist($item2);
+
         $manager->flush();
     }