+++ library/Ivoz/Provider/Application/Service/User/SyncFromCsv.php
@@ -110,6 +110,11 @@ class SyncFromCsv
                         $company,
                         ...$outboundDdiArgs
                     );
+
+                    if ($ddi->isNew()) {
+                        $user->setOutgoingDdi($ddi);
+                    }
+
                     $entities[] = $ddi;
                 }