+++ library/Ivoz/Provider/Domain/Model/Location/LocationInterface.php
@@ -1,0 +1,73 @@
+<?php
+
+namespace Ivoz\Provider\Domain\Model\Location;
+
+use Ivoz\Core\Domain\Model\LoggableEntityInterface;
+use Doctrine\Common\Collections\Criteria;
+use Doctrine\Common\Collections\ArrayCollection;
+
+interface LocationInterface extends LoggableEntityInterface
+{
+    /**
+     * @codeCoverageIgnore
+     * @return array
+     */
+    public function getChangeSet();
+
+    /**
+     * Get name
+     *
+     * @return string
+     */
+    public function getName(): string;
+
+    /**
+     * Get description
+     *
+     * @return string | null
+     */
+    public function getDescription();
+
+    /**
+     * Get company
+     *
+     * @return \Ivoz\Provider\Domain\Model\Company\CompanyInterface
+     */
+    public function getCompany();
+
+    /**
+     * @return bool
+     */
+    public function isInitialized(): bool;
+
+    /**
+     * Add user
+     *
+     * @param \Ivoz\Provider\Domain\Model\User\UserInterface $user
+     *
+     * @return static
+     */
+    public function addUser(\Ivoz\Provider\Domain\Model\User\UserInterface $user);
+
+    /**
+     * Remove user
+     *
+     * @param \Ivoz\Provider\Domain\Model\User\UserInterface $user
+     */
+    public function removeUser(\Ivoz\Provider\Domain\Model\User\UserInterface $user);
+
+    /**
+     * Replace users
+     *
+     * @param ArrayCollection $users of Ivoz\Provider\Domain\Model\User\UserInterface
+     * @return static
+     */
+    public function replaceUsers(ArrayCollection $users);
+
+    /**
+     * Get users
+     * @param Criteria | null $criteria
+     * @return \Ivoz\Provider\Domain\Model\User\UserInterface[]
+     */
+    public function getUsers(\Doctrine\Common\Collections\Criteria $criteria = null);
+}