+++ schema/DoctrineMigrations/Version20220207082635.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Application\Migrations;
+
+use Doctrine\DBAL\Migrations\AbstractMigration;
+use Doctrine\DBAL\Schema\Schema;
+
+/**
+ * Auto-generated Migration: Please modify to your needs!
+ */
+class Version20220207082635 extends AbstractMigration
+{
+    /**
+     * @param Schema $schema
+     */
+    public function up(Schema $schema)
+    {
+        $this->addSql("ALTER TABLE BillableCallHistorics ADD PRIMARY KEY(id, startTime)");
+    }
+
+    /**
+     * @param Schema $schema
+     */
+    public function down(Schema $schema)
+    {
+        $this->addSql("ALTER TABLE BillableCallHistorics DROP PRIMARY KEY");
+    }
+}