+++ web/rest/client/tests/DataAccessControl/Provider/UserTest.php
@@ -177,7 +177,21 @@ class UserTest extends KernelTestCase
                             null
                         ]
                     ]
-                ]
+                ],
+                [
+                    'or' => [
+                        [
+                            'location',
+                            'in',
+                            'LocationRepository([["company","eq","user.getCompany().getId()"]])'
+                        ],
+                        [
+                            'location',
+                            'isNull',
+                            null
+                        ]
+                    ]
+                ],
             ]
         );
     }